synapse: create a PGO DB though actually (#49)
it's a separate DB cluster, annoyingly. AFAICT there's no way to set
encoding/collation on a per DB basis. Though people have been asking for
it on PGO tracker for years.

Anyway, I expect this DB to be pretty busy, so maybe it's good that it's
separate from the "main" one?